refactor: simplify payments v2 toggle plumbing
