+++ apps/sdp-web/src/app/dashboard/payments/payments-workspace.tsx
@@ -31,6 +31,7 @@ interface PaymentsApiKeyOption {
 interface PaymentsWorkspaceProps {
   apiBaseUrl: string | null;
   apiKeys: PaymentsApiKeyOption[];
+  paymentsV2: boolean;
   wallets: PaymentsDashboardWallet[];
   walletsError: string | null;
   aggregate: CustodyWalletAggregate | null;
@@ -43,6 +44,7 @@ interface PaymentsWorkspaceProps {
 export function PaymentsWorkspace({
   apiBaseUrl,
   apiKeys,
+  paymentsV2,
   wallets,
   walletsError,
   aggregate,
@@ -51,8 +53,7 @@ export function PaymentsWorkspace({
   transfers,
   transfersError,
 }: PaymentsWorkspaceProps) {
-  const { featureFlags, issuanceTab, selectedPlaygroundApiKeyId, setPlaygroundApiKeys } =
-    useDashboardWorkspace();
+  const { issuanceTab, selectedPlaygroundApiKeyId, setPlaygroundApiKeys } = useDashboardWorkspace();
   const isPlaygroundTab = issuanceTab === "playground";
 
   useEffect(() => {
@@ -103,7 +104,7 @@ export function PaymentsWorkspace({
         <PaymentsOverview
           aggregate={aggregate}
           aggregateError={aggregateError}
-          paymentsV2={featureFlags.paymentsV2}
+          paymentsV2={paymentsV2}
           issuedTokenSymbolsByMint={issuedTokenSymbolsByMint}
           transfers={transfers}
           transfersError={transfersError}
