+++ apps/sdp-web/src/contexts/dashboard-workspace-context.tsx
@@ -18,7 +18,6 @@ import { SWRConfig } from "swr";
 import { Button } from "@/components/ui/button";
 import type { DashboardAccess } from "@/lib/dashboard-access";
 import { type DashboardCacheScope, getDashboardCacheScopeKey } from "@/lib/dashboard-cache-scope";
-import type { DashboardFeatureFlags } from "@/lib/dashboard-feature-flags";
 import { DASHBOARD_SWR_CONFIG } from "@/lib/dashboard-swr-config";
 import { useDashboardUrlState } from "@/lib/dashboard-url-state";
 import { reconcileProjectCookieAction, selectProjectAction } from "@/lib/project-cookie-action";
@@ -37,7 +36,6 @@ export interface DashboardPlaygroundApiKeyOption {
 type DashboardWorkspaceContextValue = {
   dashboardAccess: DashboardAccess;
   dashboardCacheScope: DashboardCacheScope;
-  featureFlags: DashboardFeatureFlags;
   projects: Project[];
   sandboxProject: Project | null;
   productionProject: Project | null;
@@ -81,7 +79,6 @@ type DashboardWorkspaceProviderProps = {
   children: ReactNode;
   dashboardAccess: DashboardAccess;
   serverDashboardCacheScope: DashboardCacheScope;
-  featureFlags: DashboardFeatureFlags;
   projects: Project[];
   initialSelectedProjectId: string | null;
   initialSidebarOpen?: boolean;
@@ -91,7 +88,6 @@ export function DashboardWorkspaceProvider({
   children,
   dashboardAccess,
   serverDashboardCacheScope,
-  featureFlags,
   projects,
   initialSelectedProjectId,
   initialSidebarOpen = true,
@@ -261,7 +257,6 @@ export function DashboardWorkspaceProvider({
     () => ({
       dashboardAccess,
       dashboardCacheScope: liveDashboardCacheScope,
-      featureFlags,
       projects,
       sandboxProject,
       productionProject,
@@ -284,7 +279,6 @@ export function DashboardWorkspaceProvider({
     [
       dashboardAccess,
       liveDashboardCacheScope,
-      featureFlags,
       projects,
       sandboxProject,
       productionProject,
